kp/2025/buzdugan/backend/requirements.txt
# FastAPI и связанные зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# База данных
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis для кэширования и rate limiting
redis==5.0.1
aioredis==2.0.1

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Обработка данных и приватность
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4

# Дифференциальная приватность
diffprivlib==0.6.0

# Парсинг SQL
sqlparse==0.4.4
sqlalchemy-utils==0.41.1

# Логирование и мониторинг
structlog==23.2.0
prometheus-client==0.19.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Утилиты
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2

# CORS
fastapi-cors==0.0.6

# Валидация
email-validator==2.1.0
phone-validator==0.0.1

# Криптография
cryptography==41.0.8

# Асинхронность
asyncio==3.4.3
aiohttp==3.9.1
